login as component
